refactor : detach board component
